applets/battery: Fix navigation keyboard cycle for the Up key
When on the first item of the battery list, trying to force focus
to the header item fails, probably because it is wrapped within a
Column.

Just setting it to `null` works in this case, since it automatically
completes the `KeyNavigation.down` chain, but in reverse. This is
the same trick already being used in the footer.

BUG: 489099
FIXED-IN: 6.3.1


(cherry picked from commit f7b5559a4cf88eff7d7c0f509b7e56aea5a5d2d1)

Co-authored-by: Ismael Asensio <isma.af@gmail.com>