kcms: Import the energy info module from kinfocenter
It makes sense to have the Energy InfoCenter KCM in this repo, since
it is highly related, uses its resources and even gets invoked from
the battery applet. It also shares components with the mobile version,
so they can be more easily combined in the future

This is purely a repo switch for the module's code. The KCM will
still be shown in KInfoCenter