applets/brightness: fix Return press on switch
The "Suspend" switch for Night Light calls toggle() on space/enter/
return press, and handles the onToggled() signal to actually perform the
action. The qqc2 api is exceptionally confusing here, toggle() refers to
changing the checked state of a button, but does not automatically
call toggled() which is called when a user interactively manipulates a
control, and would need to be called explicitly in the handler. The only
reason it worked for Space is that the event wasn't accepted, so the
switch called toggled() internally.

Instead, just use clicked() and avoid the whole messy toggled/toggle()
situation. Along the way, we also explicitly accept the onPressed events
when we handle them, to avoid problems with multiple activation in the
future.

BUG: 500702
FIXED-IN: 6.3.2