applets/brightness: don't skip NightLight switch on tab
The Night Light delegate has special handling to select the switch on
tab, but this is done in the delegate's contentItem, rather than
directly in the delegate. As a consequence, it only works if the
contentItem has focus, not if the delegate itself has focus, which
happens regularly.

Instead, put the handler in the delegate directly; if focus is on the
RowLayout (or elsewhere generally in the contentItem), handling will
fall back to the delegate (unless overridden) anyway so this is the only
place where we really need it.

BUG: 500703
FIXED-IN: 6.3.3