applets/brightness: downgrade severity of action unavailable message
If the KeyboardBrightnessControl action is not available, a warning is
logged. This seems a bit high, as this is the expected case for many
systems.

Instead, use Info severity - this still allows users to see the message
without also seeing all the debug messages, but removes it in the
default case (as the default minimum severity is already set to
Warning).

BUG: 501461