daemon: watch for only dpms events
If we request all events, ddcutil initializes the XEvent monitor, which
then connects with Xwayland socket provided by KWin on the Wayland
session.

Later on if kwin_wayland crashes, X connection breaks and powerdevil
safely(!) exists, and never restarts, which makes rest of power
management not work. On phones this means that you can not wake your
device up if kwin crashed when screen was off, or battery draining
because automatic suspend stuff from powerdevil was no longer running.

Also use KScreen to check for screen-reconfiguration, and if config
changes ask detectors to re-detect displays.