powerdevilcore: Unload actions when inhibited
There is no reason for us to keep these actions running when system is
inhibited. Unload them during inhibition so that their timers
stop and will not keep running.

When inhibitions are gone, check if we have any active actions,
and if not, we can then just refresh the system back to normal.

BUG: 504553