applets/batterymonitor: only show one battery icon in desktop tray
The compact representation only shows one icon in most cases, but on
the desktop users can resize it to be long in only one direction and
short in the other to get a row or column with one icon for each
connected battery. This breaks in the case of a system tray on the
desktop, where applets inherit the desktop form factor, and the battery
applet appears to have a variably-sized partial icon below it depending
on the tray widget dimensions. (I suspect the change toward nested
plasma containments in 46ada24f is involved).

Instead, we also test for the ContainmentForcesSquarePlasmoids
containment hint (which the tray sets), and consider the widget
constrained in that case - with a forced square plasmoid, showing
multiple icons doesn't make much sense anyway.

BUG: 504697
FIXED-IN: 6.4.0