daemon: Fix triggering an assert in KConfigGroup
If there is no current activity, KActivities::Consumer::currentActivity()
can return an empty string. The problem with that is that the KConfigGroup
doesn't allow empty strings to be used as group names, it has an assert
for that.

This change puts relevant code behind a guard to avoid triggering the
assert in KConfigGroup.

SENTRY: POWERDEVIL-161