applets: fix plasmashell freeze on very wide panels
When the fullRepresentation is used directly in a very wide panel,
especially if other widgets are also present, Plasma may freeze as it
gets stuck on switching between the representations - the
compactRepresentation is large enough that it wants to switch, but the
fullRepresentation ends up too small and switches back, forever.

Unsetting Layout minimum sizes (and preferredHeight) in those cases
prevents the issue from happening. This yields a square size which isn't
necessarily optimal for battery monitor with possibly lots of entries,
but better than a total freeze and is a reasonable and predictable
default size for in-panel use.