systemd: disable CAP_WAKE_ALARM
In some systems where FreeIPA, AD accounts, SSSD etc is managing users,
uid will be larger than 10000, which pam_systemd currently considers as
a container user and it will not keep CAP_WAKE_ALARM capability with
systemd-user instance. This means it will not be able to grant
CAP_WAKE_ALARM to powerdevil and it will fail systemd unit with
218/EXIT_CAPABILITIES.

This disables CAP_WAKE_ALARM in powerdevil systemd unit, only if
distribution is mobile focused, they can enable this as needed for
use-case.

BUG: 506052