suspendcontroller: snapshot wakeup count before suspending
On mobile and in some cases laptop it is quite important to know what
event triggerred wakeup from sleep. Previous commit added support for
writing /sys/power/wakeup_count, so now it is possible to identify exact
device which caused wakeup.

Snapshot the wakeup_count for each device node before we go to sleep. On
wakeup, we again take a snapshot and compare it against previously taken
snapshot.

For each changed values, we add it to list for now and print a debug
message. In upcoming commit we may decide to abort DPMS operation
because this event may not need user interaction and we can save some
power by not turning on display.