Also show popup for power cord plug/unplug
Part of https://invent.kde.org/plasma/plasma-desktop/-/issues/149

According to EU Directive 2019/882, sound cannot be the *only* method of
indicating something. Currently we do just that with the plug/unplug
notifications, though.

We have two options: turn off the sound, or show a popup as well.

This change adds a popup. I tried my best to make it not annoying in
4b0cb5e6b15d226e8f0f981c8f90530b245c5d77, so hopefully it won't drive
people crazy.