dpms/suspend: add some handling of unsoliciated wakeups
If we have wakeups from certain sources like telephony or network, often
times they do not require system to do full wakeup and once handling of
such an source is done, it is enough to suspend device again.

Add handling for this. If we get wakeups from,

- Timers (user configured alarms, systemd timers etc.)
- Telephony (3GPP events, Network change, Call, SMS)
- Network (WoWLAN packets, network change, network disconnect etc)

Currently any of these events would turn display on right away. Instead
skip DPMS::On and let consumer of event decide whether to turn display
on or not, it can be unifiedpush distributor or dialer or SMS app
checking the relevant device event and then asking power management to
do full resume.

Similarly once we get such an event, it is desirable to go back to sleep
if system was asleep previously and we don't require user attention then
we should go back to sleep ASAP.