KCM: show/hide compliance message if needed after resetting settings
Right now because the message's visibility is only manipulated by the
signal handlers that fire when the relevant comboboxes are interacted
with directly.

But when their values are changed indirectly as a result of clicking
the "Defaults" or "Reset" button, that code never runs, so the message
can remains shown or hidden when it shouldn't.

To fix this, handle the "value changed indirectly" case, too.

BUG: 507352
FIXED-IN: 6.5.0