KCM: show/hide compliance message if needed after resetting settings
Right now the message's visibility is only manipulated by the signal
handlers that fire when the relevant comboboxes are interacted with
directly.

However when their values are changed indirectly as a result of clicking
the "Defaults" or "Reset" buttons that code never runs, so the message
can remains shown or hidden when it shouldn't.

To fix this, trigger the logic any time the relevant config file key
changes, rather than only when the user explicitly interacts with the UI
elements.

BUG: 507352 
FIXED-IN: 6.5.0


(cherry picked from commit fd55c54cc10fa2096836a4805a7785a1519423cf)

49b3d85e KCM: show/hide compliance message if needed after resetting settings
f9949df6 Apply 1 suggestion(s) to 1 file(s)
93390396 Apply 1 suggestion(s) to 1 file(s)

Co-authored-by: Nate Graham <nate@kde.org>