applets/batterymonitor: Redesign inhibition list again
Highlights at a glance:
* Phrasings distinguish between sleep and screen locking
* Remove some labels, so that the user focuses on what matters

Following commit f9c81566, the PowerDevil service provides enough
information to show whether an inhibition is blocking just sleep
or just screen locking, or both.

Introducing this distinction complicated label phrasings though,
so we also had to tweak the UI to remain straightforward enough.

As per Nate's suggested design, the indicator text at the top is
removed for simplicity. The "Manually Block" is once again a toggle,
now at the top of the inhibition section.

We also get rid of the heading "N applications are currently
blocking sleep [and/or screen locking]", which used to appear
when multiple inhibitions were requested.

In order to tell whether sleep and/or screen locking are currently
inhibited, the user therefore has to look at all the remaining
individual list items. However, this is generally acceptable as more
than a small handful of inhibition list items are exceedingly rare.
(An accessible status description continues to summarize the section.)

In return, the section as a whole is tidier, generally with less
text to read, which makes it more likely that the user will actually
try to parse the remaining items.

BUG: 418433