CompactRepresentation: Don't show the badge if battery percent is 100
100% is not necessarily fully charged, but distinguishing between 100%
and fully charged and 100% and not fully charges is not that useful.
It also makes the "Show Battery Percentage on Icon When Not Fully Charged"
setting appear bugged when you unplug your laptop from the charger since
it's not obvious that 100% is not necessarily fully charged.

Some hardware (usually peripherals like mice and headphones) will
report 100% battery in situations where that is obviously not true due
to them reporting a more limited number of battery states. There's not
much we can do about those, but at least laptops and phones typically
don't do that.